Add locking buttons to the ACP
