+++ src/variables.scss
@@ -23,6 +23,11 @@
     --tavla-slider-color: #fff;
     --tavla-slider-thumb-color: #fff;
     --tavla-slider-fill-color: rgb(174, 183, 226);
+    --floating-button-background: rgb(84, 86, 140);
+    --floating-button-font-color: #fff;
+    --floating-button-primary-background: rgb(174, 183, 226);
+    --floating-button-primary-font-color: rgb(24, 28, 86);
+
 }
 
 .default-theme {
@@ -73,7 +78,10 @@
     --tavla-slider-color: #303030;
     --tavla-slider-thumb-color: #303030;
     --tavla-slider-fill-color: #4d4d4d;
-
+    --floating-button-background: rgb(42, 42, 42);
+    --floating-button-font-color: #fff;
+    --floating-button-primary-background: rgba(255, 255, 255, 0.14);
+    --floating-button-primary-font-color: #fff;
 }
 
 .light-theme {
@@ -99,6 +107,10 @@
     --tavla-slider-color: rgb(233, 233, 233);
     --tavla-slider-thumb-color: #fff;
     --tavla-slider-fill-color: rgb(148, 148, 148);
+    --floating-button-background: rgb(243, 243, 243);
+    --floating-button-font-color: #000;
+    --floating-button-primary-background: rgb(243, 243, 243);
+    --floating-button-primary-font-color: #000;
 }
 
 .grey-theme {
@@ -122,4 +134,8 @@
     --tavla-input-icon-color: #000;
     --tavla-slider-color: #e9e9e9;
     --tavla-slider-fill-color: rgb(148, 148, 148);
+    --floating-button-background: #fff;
+    --floating-button-font-color: #000;
+    --floating-button-primary-background: #fff;
+    --floating-button-primary-font-color: #000;
 }
